% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data3.R
\name{generate_data3}
\alias{generate_data3}
\title{Simulate three-group sample.}
\usage{
generate_data3(beta, gamma1, gamma2, X, G, s)
}
\arguments{
\item{beta}{coefficients of E(Y|G,X).}

\item{gamma1}{coefficients of E(Y|G,X,D=1)-E(Y|G,X,D=0)}

\item{gamma2}{coefficients of E(Y|G,X,D=2)-E(Y|G,X,D=0)}

\item{X}{Environmental Covariates. An n by q matrix (n: number of observations, q: number of covariates).}

\item{G}{Genetic covariate, valued 0,1,2 according to the allele numbers for a certain SNP. A vector with length n (n is number of observations).}

\item{s}{number of observation in the sample}
}
\value{
data: The generated sample, containing secondary trait y, primary trait D, G and X
}
\description{
\code{generate_data3} Simulate the whole population data containing 100000 observations with certain subgroup prevalence and generate a three-group sample drawn from the whole set with group proportions (25%, 50%, 25%). The output will be a three-grouop sample for the following analysis. This simulation procedure is used in our paper.
}
\examples{

#suppose the size of the whole population is 100000 with three subgruops and the proportions of the three groups are (10\%, 15\%, 75\%)
# Generate genetic covariate G with MAF=0.3
MAF <- 0.3
G <- rbinom(100000,2,MAF)
s <- 800
beta <- c(5, 2, 1)
gamma1 <- c(-2,-1,-1)
gamma2 <- c(1,1,1)
# Generate X, which is mixture of normal variables
x_p <- c(0.2,0.8)
x_mean <- c(0,2)
x_1 <- rnorm(20000, mean = 2, sd = 2)
x_2 <- rnorm(80000, mean=0, sd=2)
X <- c(x_1,x_2)
data <- generate_data3(beta, gamma1, gamma2, X, G, s)

}
\references{
Fan Zhou, Haibo Zhou, Tengfei Li, Hongtu Zhu. "Analysis of Secondary Phenotypes in Multiple-group Association Studies"
}
\author{
Fan Zhou
}
