\name{MGLREG-package}
\docType{package}
\title{
Multiple-Group sample based Linear Regression model
}
\description{
The aim of this package is to develop a general regression framework for the analysis of secondary phenotypes collected in multiple group association studies. Our regression framework is built on a conditional model for the secondary outcome given the multiple group status and covariates and its relationship with the population regression of interest of the secondary outcome given covariates. Then, we develop generalized estimating equations to estimate parameters of interest. We use simulations and a large-scale imaging genetic data analysis of the ADNI data to evaluate the effects of the multiple group sampling scheme on genome-wide association study (GWAS) results based on some standard statistical methods, such as linear regression methods, while comparing it with our statistical methods that appropriately adjust for the multiple group sampling scheme.
}
\details{
\tabular{ll}{
Package: \tab STAMG\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2017-10-18\cr
License: \tab GPL-2 | GPL-3\cr
}
}
\author{
Fan Zhou, Hongtu Zhu\cr
Maintainer: Fan Zhou <mlzxzhou@live.unc.edu>
}
\references{
Fan Zhou, Haibo Zhou, Tengfei Li, Hongtu Zhu. "Analysis of Secondary Phenotypes in Multiple-group Association Studies".
}
