% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGLReg.R
\name{MGLReg}
\alias{MGLReg}
\title{MGLReg on multiple-group sample (single-phase).}
\usage{
MGLReg(y, G, X, D, p_til)
}
\arguments{
\item{y}{Response. Currently, we only consider continuous variables. A vector with length n ( n is number of observations).}

\item{G}{Genetic covariate, valued 0,1,2 according to the allele numbers for a certain SNP. A vector with length n (n is number of observations).}

\item{X}{Environmental Covariates such as age, sex and first few PCs to adjust for strafication. An n by q matrix (n: number of observations, q: number of covariates).}

\item{D}{Primary status. Valued 0,1,2 in the three-subgroup case, representing the belongs of obervation to a certain subgroup.}

\item{p_til}{Subgroup prevalence in the whole population. A vector with length 3 in the three-subgroup case.}
}
\value{
A list object.

coef: Estimation of coefficient of genetic covariate G in the regression model given by our method.

        pvalue : p-value of the corresponding genetic effect
}
\description{
\code{STACPreg3} This function is to deal with the basic single-phase three-group case discussed in the paper by running the unbiased linear regression model MGLReg. The output includes the estimation, asymptotical variance and the corresponding p-value of the genetic effect.
}
\examples{
# suppose the size of the whole population is 100000 with three subgruops and the proportions of the three groups are (10\%, 15\%, 75\%)
# Generate genetic covariate G with MAF=0.2
MAF <- 0.3
G <- rbinom(100000,2,MAF)
s <- 800
beta <- c(5, 2, 1)
gamma1 <- c(-2,-1,-1)
gamma2 <- c(1,1,1)
# Generate X, which is mixture of normal variables
x_p <- c(0.2,0.8)
x_mean <- c(0,2)
x_1 <- rnorm(20000, mean = 2, sd = 2)
x_2 <- rnorm(80000, mean=0, sd=2)
X <- c(x_1,x_2)
data <- generate_data3(beta, gamma1, gamma2, X, G, s)
reg <- STACPreg3(data[,2],data[,3],data[,4],data[,1],c(0.1,0.15,0.75))
bias <- abs(reg[[1]]-beta[2]) ## the bias by using our approach
}
\references{
Fan Zhou, Haibo Zhou, Tengfei Li, Hongtu Zhu. "Analysis of Secondary Phenotypes in Multiple-group Association Studies"
}
\author{
Fan Zhou
}
