% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGLReg2.R
\name{MGLReg2}
\alias{MGLReg2}
\title{MGLReg on multiple-group sample (two-phase).}
\usage{
MGLReg2(y, G, C, D, p, p_til)
}
\arguments{
\item{y}{Response. Currently, we only consider continuous variables. A vector with length n ( n is number of observations).}

\item{G}{Genetic covariate, valued 0,1,2 according to the allele numbers for a certain SNP. A vector with length n (n is number of observations).}

\item{C}{Environmental Covariates such as age, sex and first few PCs to adjust for strafication. An n by q matrix (n: number of observations, q: number of covariates).}

\item{D}{Primary status. Valued 0,1,2 in the three-subgroup case, representing the belongs of obervation to a certain subgroup.}

\item{p}{Phase status, value 0 for the first phase, 1 for the second phase. We assume that all the subjects are sorted by phase number.}

\item{p_til}{Subgroup prevalence in the whole population. A vector with length 3 in the three-subgroup case.}
}
\value{
A list object.

coef: Estimation of coefficient of genetic covariate G in the regression model given by our method.

        pvalue : p-value of the corresponding genetic effect
}
\description{
\code{STACPreg3} This function is to deal with the two-phase case discussed in the paper by running the unbiased linear regression model MGLReg. The output includes the estimation, asymptotical variance and the corresponding p-value of the genetic effect.
}
\examples{
# Gwas study of SNPs from Chr22 on the volume of right hippocampus
p_til=c(0.8,0.1,0.1) # the pre-set global prevelance of three groups
data('real_data')
total_snp=dim(Geno)[2]
p_value=matrix(0,total_snp,2)
p_value[,1]=colnames(Geno)
for (i in 1:total_snp){
 p_value[i,2]=MGLReg2(y,Geno[,i],C,D,p,p_til)[[2]]
}
}
\references{
Fan Zhou, Haibo Zhou, Tengfei Li, Hongtu Zhu. "Analysis of Secondary Phenotypes in Multiple-group Association Studies"
}
\author{
Fan Zhou
}
